# Performance: only trigger dropzone handling when a user intends to drop an item