--- a/packages/e2e-tests/specs/editor/various/list-view.test.js
+++ b/packages/e2e-tests/specs/editor/various/list-view.test.js
@@ -16,7 +16,9 @@ async function dragAndDrop( draggableElement, targetElement, offsetY ) {
 		y: targetClickablePoint.y + offsetY,
 	};
 
-	return await page.mouse.dragAndDrop( draggablePoint, targetPoint );
+	return await page.mouse.dragAndDrop( draggablePoint, targetPoint, {
+		delay: 100,
+	} );
 }
 
 describe( 'List view', () => {