### Background 

Our drop zone handlers are also causing major performance issues when dragging quickly. We have dragging listeners on every component that is a drop zone. These events are currently not throttled and when dragging quickly we can queue up quite a number of events to resolve. We will also trigger a number of unneeded animations. This slows things down quite a bit.

In this PR we try and infer what the user intends to do. Are they trying to move the item quickly, someplace else or are they ready to drop an item into a drop zone? If they're not ready to drop, we can skip work.

I profiled using the Armando theme, and the "alternating rows with images and headings" pattern.

| before | after |
|-----|-----|
| <img width="1422" alt="before" src="https://user-images.githubusercontent.com/1270189/126818329-7a32d9bc-0794-40db-a5a0-1857268c22f1.png"> | <img width="1263" alt="after-16" src="https://user-images.githubusercontent.com/1270189/127193155-e22090dd-603b-45ba-8bf0-ce91f5bc106c.png"> |

Here are some videos of the type of actions I was performing while profiling. Note that screenrecording + profiling also impacts performance, so I took separate profiling screencaps above.

#### Profiling In Trunk 

https://user-images.githubusercontent.com/1270189/126814474-11aeca2e-bbe0-407a-a075-353506b0a74f.mp4

#### Branch 

https://user-images.githubusercontent.com/1270189/127194167-b6eae4f8-3858-46db-b084-e9e7223f4e9b.mp4

### Testing Instructions

- No regressions in dropzone
- Verify that the drop indicators feel responsive enough. (We can still tune this further if we're sampling not often enough.)

#### Follow Up

To improve drag performance further I think there's two areas of interest I've spotted:
- Incremental follow up of exploring if dragleave events can be throttled further.
- Invert control, and explore a global drag handler instead of many child listeners trying to cancel each other's events. Much larger refactor, but I'd typically pick this setup for a complex editor/app.