@gwwar 2021-07-16
I did a little **side exploration** (with extremely rough code), to see what would happen if we added one set of **global drag handlers to the document**, instead of per drop zone in https://github.com/WordPress/gutenberg/commit/9eee3045789ea76073f6c4dd7cee36ce62a2ace8

**I'm going to defer on this approach for now**, since the chance of regression is a bit too high, but I think it might be worth revisiting if we're still seeing some bad performance characteristics with drag in the future.

A few things I discovered:
- Pretty good potential performance improvement. 
- In general I find a single global handler usually helps simplify things, especially when only one handler should fire on drop (we also avoid needing to cancel the right events). However it probably goes against our general code style.
- The drop event is a bit quirky. In order for the actual drop event to fire, we should `event.preventDefault()` on the dragover or dragenter handler. https://developer.mozilla.org/en-US/docs/Web/API/HTML_Drag_and_Drop_API/Drag_operations#drop https://developer.mozilla.org/en-US/docs/Web/API/Document/drag_event
- This is workable I think, but might high a high risk of regression. We'd need to rework some callbacks to remove some assumptions about the drop lifecycle

https://user-images.githubusercontent.com/1270189/126010247-8dac2039-4380-4a9b-981e-997b505f1aa2.mp4