@gwwar 2021-07-16
Ran out of time, but I did a very basic distance check in https://github.com/WordPress/gutenberg/pull/33435/commits/4adbcf79e99a69642be450aa395dd2c68f369b5e which looks promising.

Things to investigate next week would be:
- See if we can throttle dragover and cancel the dragenter event instead, to make sure the drop event is available.
- Make sure we're calculating relative distance, taking into account the time passed from the last frame. We'll need this to better tune when insertion hints show up.

https://user-images.githubusercontent.com/1270189/126012857-a942e06d-07cf-42ed-949f-e503a11674e4.mp4