@gwwar 2021-07-23
> List view allows a user to drag a block to a new sibling position

Hmm, looks like that's unhappy in CI, but works locally for me. I'll try and see if other tuning options work out a bit better.