@gwwar 2021-07-23
Updated the summary with the new profiling timings with the 33ms sampling.