@gwwar 2021-07-27
@talldan much appreciated if you could see if drag performance feels as responsive (or better) in the persistent list view.