@gwwar 2021-07-28
> There might be some interesting experimentation here - like when the user is dragging quickly the indicator could just follow the cursor (a simple calculation), and only when the cursor settles in a position could it snap to the right block position (the complex calculation). I don't know exactly how that would feel but it'd be interesting to try.

🤔 @talldan One thing we can try first is provide better drag hints (vs the 1px line) in a separate PR. This would give us better UX and would make this a bit more obvious as we try and tune drag performance a bit.

> This is one of those interesting cases where UX intersects performance. 

💯 For example, I'm quite aware of how slow/sluggish that drag chip is on trunk, to the point where I normally don't try to perform any editing actions with drag. At the same time it'll be pretty frustrating for folks if dropping items into an area isn't responsive enough (eg we had a false negative that they were trying to drag an item somewhere else quickly.)

> A new loop is only triggered when the previous one has completed. I think that could manifest itself as more of a locking mechanism

Work/render loops can be quite useful in complicated apps. It's possible for us to also use [requestAnimationFrame](https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame) here if we like results. (As an aside a single global drag handler that calls the right child callbacks is much faster, but also is a big reactor).