@gwwar 2021-08-11
Going to close this one out for now in favor of some new drag interaction experiments